Move runner test function from main package
